+++ lib/src/anbocas_flutter_ticket_booking.dart
@@ -1,7 +1,9 @@
 import 'package:anbocas_tickets_ui/anbocas_tickets_ui.dart';
 import 'package:anbocas_tickets_api/anbocas_tickets_api.dart';
 import 'package:anbocas_tickets_ui/src/helper/size_utils.dart';
+import 'package:anbocas_tickets_ui/src/screens/manage_attendies/event_check_in_list_screen.dart';
 import 'package:anbocas_tickets_ui/src/screens/ticket_crud/ticket_listing_screen.dart';
+import 'package:anbocas_tickets_ui/src/screens/ticket_purchase/anbocas_order_detail_screen.dart';
 import 'package:anbocas_tickets_ui/src/service/anbocas_booking_manager.dart';
 import 'package:flutter/material.dart';
 
@@ -14,36 +16,42 @@ class AnbocasTickets {
 
   String sandboxUrl = 'https://sandbox-api.anbocas.com';
   String productionUrl = 'https://api.anbocas.com';
-  late String apiUrl;
+  String baseUrl = 'https://api.anbocas.com';
+
+  String? anbocasRazorpayApiKey;
 
   void config({
     required String apikey,
-    AnbocasCustomTheme? customThemeConfig,
     ApiMode apiMode = ApiMode.sandbox,
+    AnbocasCustomTheme? customThemeConfig,
+    required String anbocasRazorpayApiKey,
   }) {
     final serviceManager = AnbocasServiceManager();
-    if (apiMode == ApiMode.production) {
-      apiUrl = productionUrl;
+    this.anbocasRazorpayApiKey = anbocasRazorpayApiKey;
+
+    if (apiMode == ApiMode.sandbox) {
+      baseUrl = 'https://sandbox-api.anbocas.com';
     } else {
-      apiUrl = sandboxUrl;
+      baseUrl = 'https://api.anbocas.com';
     }
-    serviceManager.initializeBookingRepo(apiUrl, apikey);
+
+    serviceManager.initializeBookingRepo(baseUrl, apikey);
     theme.updateConfig(customThemeConfig);
-    AnbocasTicketsApi.instance?.config(token: apikey, enableLog: true);
+    AnbocasTicketsApi.instance
+        ?.config(token: apikey, enableLog: true, mode: apiMode);
   }
 
   void launchBookingFlow({
     required String eventId,
+    String? referenceEventId,
     required BuildContext context,
-    bool allowGroupTicket = false,
     UserConfig? userMetaData,
   }) {
     try {
       if (userMetaData != null) {
         userConfig.updateConfig(userMetaData);
       }
-      Navigator.push(
-        context,
+      Navigator.of(context, rootNavigator: true).push(
         PageRouteBuilder(
           pageBuilder: (ctx, __, ___) {
             if (MediaQueryHolder().mediaQueryData == null) {
@@ -51,7 +59,7 @@ class AnbocasTickets {
             }
             return AnbocasTicketBookingWidget(
               eventId: eventId,
-              allowGroupTicket: allowGroupTicket,
+              referenceEventId: referenceEventId,
             );
           },
         ),
@@ -66,8 +74,7 @@ class AnbocasTickets {
     required String eventId,
   }) {
     try {
-      Navigator.push(
-        context,
+      Navigator.of(context, rootNavigator: true).push(
         PageRouteBuilder(
           pageBuilder: (ctx, __, ___) {
             if (MediaQueryHolder().mediaQueryData == null) {
@@ -81,6 +88,52 @@ class AnbocasTickets {
       );
     } catch (e) {
       // log(e.toString());
+    }
+  }
+
+  void manageAttendees({
+    required BuildContext context,
+    required String eventId,
+  }) {
+    try {
+      Navigator.of(context, rootNavigator: true).push(
+        PageRouteBuilder(
+          pageBuilder: (ctx, __, ___) {
+            if (MediaQueryHolder().mediaQueryData == null) {
+              MediaQueryHolder().initialize(ctx);
+            }
+            return EventCheckInListScreen(
+              eventId: eventId,
+            );
+          },
+        ),
+      );
+    } catch (e) {
+      // log(e.toString());
+    }
+  }
+
+  void viewOrderSummary({
+    required BuildContext context,
+    required String anbocasOrderId,
+    String? referenceEventId,
+  }) {
+    try {
+      Navigator.of(context, rootNavigator: true).push(
+        PageRouteBuilder(
+          pageBuilder: (ctx, __, ___) {
+            if (MediaQueryHolder().mediaQueryData == null) {
+              MediaQueryHolder().initialize(ctx);
+            }
+            return AnbocasOrderDetailScreen(
+              anbocasOrderId: anbocasOrderId,
+              referenceEventId: referenceEventId,
+            );
+          },
+        ),
+      );
+    } catch (e) {
+      debugPrint(e.toString());
     }
   }
 }