+++ lib/src/components/ticket_single_item.dart
@@ -1,33 +1,32 @@
 import 'package:anbocas_tickets_ui/anbocas_tickets_ui.dart';
-import 'package:anbocas_tickets_ui/src/components/custom_button.dart';
+import 'package:anbocas_tickets_ui/src/anbocas_flutter_ticket_booking.dart';
 import 'package:anbocas_tickets_ui/src/components/dottted_line.dart';
 import 'package:anbocas_tickets_ui/src/components/icon_with_circle_background.dart';
 import 'package:anbocas_tickets_ui/src/components/ticket_card_clipper.dart';
 import 'package:anbocas_tickets_ui/src/helper/size_utils.dart';
 import 'package:anbocas_tickets_ui/src/model/single_ticket.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_widget_from_html/flutter_widget_from_html.dart';
 
 class TicketItemWidget extends StatefulWidget {
-  const TicketItemWidget(
-      {super.key,
-      required this.element,
-      this.isSelected = false,
-      this.showBuyButton = false,
-      this.showDeleteIcon = false,
-      required this.onQuantityChanged,
-      this.onDeletePres,
-      this.onBuyItemPressed,
-      this.onItemSelect});
+  const TicketItemWidget({
+    super.key,
+    required this.element,
+    this.isSelected = false,
+    this.showBuyButton = false,
+    this.showDeleteIcon = false,
+    required this.onQuantityChanged,
+    this.onDeletePres,
+    this.onItemSelect,
+  });
 
-  final SingleTickets element;
+  final SingleTicket element;
   final bool isSelected;
   final bool showBuyButton;
   final bool showDeleteIcon;
-  final ValueChanged<int> onQuantityChanged;
+  final void Function(int newQuantity, String ticketId) onQuantityChanged;
   final void Function()? onDeletePres;
-  final void Function()? onBuyItemPressed;
   final void Function()? onItemSelect;
+
   @override
   State<TicketItemWidget> createState() => _TicketItemWidgetState();
 }
@@ -44,35 +43,42 @@ class _TicketItemWidgetState extends State<TicketItemWidget> {
 
   @override
   void didUpdateWidget(covariant TicketItemWidget oldWidget) {
-    if (oldWidget.isSelected != widget.isSelected) {
-      quantity.value = 1;
-      widget.onQuantityChanged(quantity.value);
-    }
-    if (oldWidget.element != widget.element) {
-      quantity.value = widget.element.selectedQuantity;
-      widget.onQuantityChanged(quantity.value);
-    }
+    // if (oldWidget.isSelected != widget.isSelected) {
+    //   quantity.value = 1;
+    //   widget.onQuantityChanged(quantity.value);
+    // }
+    // if (oldWidget.element != widget.element) {
+    //   quantity.value = widget.element.selectedQuantity;
+    //   widget.onQuantityChanged(quantity.value);
+    // }
     super.didUpdateWidget(oldWidget);
   }
 
   @override
   Widget build(BuildContext context) {
     return Container(
-      height: (widget.isSelected || widget.showBuyButton) ? 180.v : 140.v,
+      margin: EdgeInsets.only(bottom: 10.h),
+      height: 160.v,
       width: double.infinity,
       decoration: BoxDecoration(
-          color: widget.isSelected == true ? theme.accentColor : null),
+          color: widget.isSelected == true
+              ? theme.ticketCardConfig.selectedTicketCardBorderColor
+              : null),
       child: CustomPaint(
         painter: TicketCardFillPainter(radius: 15, isSelected: true),
         child: ClipPath(
           clipper: TicketCardClipper(radius: 15),
           child: DecoratedBox(
             decoration: BoxDecoration(
-                borderRadius: BorderRadius.circular(10),
-                border: widget.isSelected == true
-                    ? Border.all(color: theme.accentColor!, width: 2.adaptSize)
-                    : const Border(),
-                color: theme.secondaryBgColor),
+              borderRadius: BorderRadius.circular(10),
+              border: widget.isSelected == true
+                  ? Border.all(
+                      color:
+                          theme.ticketCardConfig.selectedTicketCardBorderColor,
+                      width: 2.adaptSize)
+                  : const Border(),
+              color: theme.ticketCardConfig.ticketCardBackgroundColor,
+            ),
             child:
                 Column(crossAxisAlignment: CrossAxisAlignment.start, children: [
               Expanded(
@@ -90,128 +96,100 @@ class _TicketItemWidgetState extends State<TicketItemWidget> {
                           crossAxisAlignment: CrossAxisAlignment.start,
                           children: [
                             Expanded(
-                              child: Text(
-                                widget.element.name ?? "",
-                                maxLines: 2,
-                                overflow: TextOverflow.ellipsis,
-                                style: theme.bodyStyle?.copyWith(
-                                  fontWeight: FontWeight.w700,
-                                  fontSize: 16.fSize,
-                                ),
-                              ),
+                              child: Text(widget.element.name ?? "",
+                                  maxLines: 2,
+                                  overflow: TextOverflow.ellipsis,
+                                  style: theme.ticketCardConfig.nameStyle),
                             ),
                             SizedBox(
                               width: 20.h,
                             ),
                             Text(widget.element.formattedPrice ?? "",
-                                style: theme.bodyStyle)
+                                style: theme.ticketCardConfig.priceStyle),
                           ],
                         ),
                         SizedBox(
                           height: 10.v,
                         ),
-                        HtmlWidget(
+                        Text(
                           widget.element.description ?? "",
-                          textStyle: theme.smallLabelStyle,
+                          style: theme.ticketCardConfig.labelStyle
+                              .copyWith(overflow: TextOverflow.ellipsis),
+                          maxLines: 2,
                         ),
                       ],
                     ),
                   ),
                 ),
               ),
-              if (widget.isSelected || widget.showBuyButton)
-                CustomPaint(
-                  painter: DottedLine(radius: 20),
-                  child: const SizedBox(
-                    width: double.infinity,
-                  ),
+              CustomPaint(
+                painter: DottedLine(
+                    radius: 20, color: theme.ticketCardConfig.dottedLineColor),
+                child: const SizedBox(
+                  width: double.infinity,
                 ),
-              if (widget.isSelected || widget.showBuyButton)
-                Expanded(
-                    child: Padding(
-                  padding:
-                      EdgeInsets.symmetric(vertical: 15.v, horizontal: 20.h),
-                  child: Row(
-                    crossAxisAlignment: CrossAxisAlignment.center,
-                    children: [
-                      if (widget.showBuyButton)
-                        Padding(
-                          padding: EdgeInsets.only(right: 10.0.h),
-                          child: CustomButton(
-                            onPressedCallback: widget.onBuyItemPressed,
-                            centerText: "Buy Ticket",
-                            buttonSize: Size(60.h, 40.v),
+              ),
+              Padding(
+                padding: EdgeInsets.symmetric(vertical: 15.v, horizontal: 20.h),
+                child: (widget.element.available != 0)
+                    ? Row(
+                        mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                        children: [
+                          Text(
+                            "Quantity",
+                            style: theme.ticketCardConfig.labelStyle,
                           ),
-                        ),
-                      if (!widget.showBuyButton)
-                        Row(
-                          children: [
-                            Text(
-                              "Quantity",
-                              style: theme.bodyStyle?.copyWith(
-                                fontWeight: FontWeight.w500,
-                                fontSize: 15.fSize,
-                                color: theme.secondaryTextColor,
+                          SizedBox(
+                            width: 10.h,
+                          ),
+                          Row(
+                            children: [
+                              IconWithCircleBackground(
+                                  icon: Icons.remove,
+                                  onPressed: () {
+                                    if (quantity.value > 0) {
+                                      quantity.value--;
+                                      widget.onQuantityChanged(
+                                          quantity.value, widget.element.id!);
+                                    }
+                                  },
+                                  color: theme.ticketCardConfig
+                                      .qtyReduceBackgroundColor),
+                              SizedBox(
+                                width: 10.h,
                               ),
-                            ),
-                            SizedBox(
-                              width: 10.h,
-                            ),
-                            // const Spacer(),
-                            IconWithCircleBackground(
-                                icon: Icons.remove,
-                                onPressed: () {
-                                  if (quantity.value >
-                                          widget.element.minQtyPerOrder &&
-                                      widget.isSelected == true) {
-                                    quantity.value--;
-                                    widget.onQuantityChanged(quantity.value);
-                                  }
-                                },
-                                color:
-                                    theme.secondaryBgColor!.withOpacity(0.9)),
-                            SizedBox(
-                              width: 10.h,
-                            ),
-                            ValueListenableBuilder<int>(
-                                valueListenable: quantity,
-                                builder: (context, quantity, child) {
-                                  return Text(
-                                    quantity.toString(),
-                                    style: theme.labelStyle?.copyWith(
-                                      color: theme.primaryTextColor,
-                                    ),
-                                  );
-                                }),
-                            SizedBox(
-                              width: 10.h,
-                            ),
-                            IconWithCircleBackground(
-                                onPressed: () {
-                                  if (widget.element.maxQtyPerOrder >
-                                          quantity.value &&
-                                      widget.isSelected == true) {
-                                    quantity.value++;
-                                    widget.onQuantityChanged(quantity.value);
-                                  }
-                                },
-                                icon: Icons.add,
-                                color: theme.secondaryIconColor!),
-                          ],
-                        ),
-                      if (widget.showDeleteIcon) const Spacer(),
-                      if (widget.showDeleteIcon)
-                        IconButton(
-                            onPressed: widget.onDeletePres,
-                            alignment: Alignment.center,
-                            padding: EdgeInsets.zero,
-                            icon: Icon(
-                              Icons.delete,
-                              color: theme.secondaryIconColor,
-                            ))
-                    ],
-                  ),
-                ))
+                              ValueListenableBuilder<int>(
+                                  valueListenable: quantity,
+                                  builder: (context, quantity, child) {
+                                    return Text(
+                                      quantity.toString(),
+                                      style: theme.labelStyle?.copyWith(
+                                        color: theme.primaryTextColor,
+                                      ),
+                                    );
+                                  }),
+                              SizedBox(
+                                width: 10.h,
+                              ),
+                              IconWithCircleBackground(
+                                  onPressed: () {
+                                    if (quantity.value <= 9) {
+                                      quantity.value++;
+                                      widget.onQuantityChanged(
+                                          quantity.value, widget.element.id!);
+                                    }
+                                  },
+                                  icon: Icons.add,
+                                  color: theme
+                                      .ticketCardConfig.qtyAddBackgroundColor),
+                            ],
+                          )
+                        ],
+                      )
+                    : Text("Sold Out!",
+                        style: theme.ticketCardConfig.labelStyle
+                            .copyWith(color: theme.errorColor)),
+              ),
             ]),
           ),
         ),