+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+## 0.0.5
+
+* Navigation Fixes
+* Added View Order Detail Screen
+* UI Fixes
+
+## 0.0.4
+
+* Bug Fixes for event not fired when View Event was clicked on the Order Detail Screen
+
+## 0.0.3
+
+* Bug Fixes
+* Minor UI Enhancements
+
+## 0.0.2
+
+* Code Improvements
+* UI Enhancements
+* Code Cleanup
+
 ## 0.0.1
 
-* Describe initial release.
+* Inital Release of the plugin.